tests/service/requirements.txt
requests
azure-mgmt-resource==0.30.0rc5
keyring
sev-snp-measure